requirements.txt
# Interfaz de usuario
streamlit>=1.28.0

# Librerías para extracción de PDFs
pdfplumber>=0.10.0
PyPDF2>=3.0.0
camelot-py[cv]>=0.10.0
tabula-py>=2.8.0
PyMuPDF>=1.23.0

# Librerías para procesamiento de datos
pandas>=2.0.0
numpy>=1.24.0
openpyxl>=3.1.0

# Integraciones Google (opcional - para modo colaborativo)
google-api-python-client>=2.100.0
google-auth-oauthlib>=1.2.0
google-auth>=2.20.0
google-auth-httplib2>=0.2.0
gspread>=5.10.0

# Librerías para análisis y visualización
matplotlib>=3.7.0
seaborn>=0.12.0

# Utilidades
python-dotenv>=1.0.0
tqdm>=4.65.0

# Para desarrollo y testing
pytest>=7.0.0
black>=23.0.0

